[REFACTOR] 주문상태 숨김처리 추가
