+++ src/components/mypage/MyGroupOrders.tsx
@@ -1,29 +1,29 @@
 import { useState, useEffect } from 'react';
-import { fetchMyNormalOrders, fetchMyPageSummary } from '@/api/Mypage';
+import { fetchMyGroupOrders, fetchMyPageSummary } from '@/api/Mypage';
 import { MyPageSummary } from '@/types/Mypage';
 
-import styles from '@/css/main/MyNormalOrders.module.css';
+import styles from '@/css/main/MyGroupOrders.module.css';
 
 import OrderHistoryList from '@/components/mypage/OrderHistoryList';
 import OrderSummaryStatus from '@/components/mypage/OrderSummaryStatus';
 import Spinner from '@/components/common/Spinner';
 
-function MyNormalOrders() {
+function MyGroupOrders () {
     const [userSummary, setUserSummary] = useState<MyPageSummary | null>(null); // 사용자 정보
-    const [normalOrder, setNormalOrder] = useState([]); // 공동 주문 내역
+    const [groupOrder, setGroupOrder] = useState([]);   // 공동 주문 내역
 
     useEffect(() => {
         const loadData = async () => {
             try {
                 const [summaryData, orderData] = await Promise.all([
                     fetchMyPageSummary(),
-                    fetchMyNormalOrders()
+                    fetchMyGroupOrders()
                 ]);
 
-                setNormalOrder(orderData.orders);
+                setGroupOrder(orderData.orders);
                 setUserSummary(summaryData);
             } catch (error) {
-                console.error("일반 주문 데이터 불러오기 실패:", error);
+                console.error("공동 주문 데이터 불러오기 실패:", error);
             }
         };
 
@@ -33,16 +33,16 @@ function MyNormalOrders() {
     // 주문 취소 버튼 클릭시 실행
     const handleOrderCancel = (orderId: number) => {
         // TODO: 여기서 주문 취소로 이어져야합니다.
-        setNormalOrder(prev => 
-            prev.map(group => 
-                group.orderId === orderId 
-                    ? { 
-                        ...group, 
-                        canCancel: false, 
-                        items: group.items.map(item => ({ 
-                            ...item, 
+        setGroupOrder(prev =>
+            prev.map(group =>
+                group.orderId === orderId
+                    ? {
+                        ...group,
+                        canCancel: false,
+                        items: group.items.map(item => ({
+                            ...item,
                             orderStatus: '주문 취소'
-                        })) 
+                        }))
                     }
                     : group
             )
@@ -59,7 +59,7 @@ function MyNormalOrders() {
     }
 
     return (
-        <div className={styles.normalOrdersPage}>
+        <div className={styles.groupOrdersPage}>
             {/* 사용자 정보 및 주문 통계 컴포넌트 */}
             <OrderSummaryStatus
                 userName={userSummary.name}
@@ -68,19 +68,20 @@ function MyNormalOrders() {
                 groupOrderAmount={userSummary.groupOrderAmount}
                 customDosirakAmount={userSummary.customDosirakAmount}
             />
-            
+
             {/* 섹션 제목 */}
-            <h1>일반 주문 내역</h1>
-            
-            {/* 일반주문 내역 목록 컴포넌트 */}
+            <h1>공동 주문 내역</h1>
+
+            {/* 공동주문 내역 목록 컴포넌트 */}
             <OrderHistoryList
-                orderGroups={normalOrder}
+                orderGroups={groupOrder}
                 onCancelClick={handleOrderCancel}
                 showMoreButton={false}
                 hideOrderHeader={false}
+                hideStatusBadge={false}
             />
         </div>
     );
 };
 
-export default MyNormalOrders;
+export default MyGroupOrders;
